dev iei

See merge request qa-produit/tests/qa_tina_tests!781

+++ qa_tina_tests/USER/API/OAPI/Vm/test_CreateVms.py
@@ -1,4 +1,3 @@
-
 from __future__ import division
 
 import base64
@@ -1037,6 +1036,7 @@ class Test_CreateVmsWithSubnet(OscTinaTest):
     def test_T4401_with_two_valid_nic_id(self):
         nic_id2 = None
         subnet_id2 = None
+        ret = None
         try:
             subnet_id2 = self.a1_r1.oapi.CreateSubnet(NetId=self.net_id,
                                                       IpRange='10.1.1.0/24').response.Subnet.SubnetId
@@ -1061,7 +1061,7 @@ class Test_CreateVmsWithSubnet(OscTinaTest):
                 }],
             )
         finally:
-            if nic_id2:
+            if nic_id2 and ret:
                 self.a1_r1.oapi.UnlinkNic(LinkNicId=ret.response.Vms[0].Nics[1].LinkNic.LinkNicId)
                 wait_network_interfaces_state(osc_sdk=self.a1_r1, network_interface_id_list=[nic_id2], state='available')
                 self.a1_r1.oapi.DeleteNic(NicId=nic_id2)
